ci: test bloaty PR

+++ rmk/src/lib.rs
@@ -94,7 +94,7 @@ pub mod storage;
 pub(crate) mod usb;
 pub mod via;
 
-const TMP: usize = 1;
+const TMP: u8 = 1;
 
 pub async fn initialize_keymap<const ROW: usize, const COL: usize, const NUM_LAYER: usize>(
     default_keymap: &mut [[[action::KeyAction; COL]; ROW]; NUM_LAYER],
